ADD: Refactorizacion entities
