+++ src/main/java/com/example/forge/auth/SimpleGrantedAuthorityJsonCreator.java
@@ -5,7 +5,5 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public abstract class SimpleGrantedAuthorityJsonCreator {
   @JsonCreator
-  public SimpleGrantedAuthorityJsonCreator(@JsonProperty("authority") String role) {
-
-  }
+  public SimpleGrantedAuthorityJsonCreator(@JsonProperty("authority") String role) { }
 }
