+++ src/main/java/com/example/forge/models/BaseEntity.java
@@ -1,7 +1,9 @@
 package com.example.forge.models;
 
 import java.util.Date;
+
 import org.springframework.format.annotation.DateTimeFormat;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -12,27 +14,27 @@ import jakarta.persistence.PreUpdate;
 import lombok.Getter;
 import lombok.Setter;
 
-@MappedSuperclass
 @Getter
 @Setter
+@MappedSuperclass
 public abstract class BaseEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
-  
+
   @Column(updatable=false, name = "created_at")
   @DateTimeFormat(pattern="yyyy-MM-dd")
   private Date createdAt;
-    
+
   @Column(name = "updated_at")
   @DateTimeFormat(pattern="yyyy-MM-dd")
   private Date updatedAt;
-    
+
   @PrePersist
   protected void onCreate(){
 	this.createdAt = new Date();
   }
-    
+
   @PreUpdate
   protected void onUpdate(){
 	this.updatedAt = new Date();
