+++ src/main/java/com/example/forge/models/entities/Administration.java
@@ -8,24 +8,21 @@ import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@NoArgsConstructor
-@AllArgsConstructor
 @Table(name="administrations")
 @Getter
 @Setter
+@NoArgsConstructor
 public class Administration extends BaseEntity {
-	
 	@NotNull
 	@NotBlank
 	@Size(min = 4, max = 15)
 	private String name;
-	
+
 	@NotNull
 	@NotBlank
 	@Size(min = 4, max = 15)
