+++ src/main/java/com/example/forge/models/entities/Degree.java
@@ -8,30 +8,27 @@ import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@NoArgsConstructor
-@AllArgsConstructor
 @Table(name="degrees")
 @Getter
 @Setter
+@NoArgsConstructor
 public class Degree extends BaseEntity {
-
 	@NotNull
 	@NotBlank
 	@Size(min = 4, max = 15)
 	private String name;
-	
+
 	@NotNull
 	@NotBlank
 	@Size(min = 4, max = 15)
 	@Column(name = "assigned_room")
 	private String assignedRoom;
-	
+
 	@NotNull
 	@NotBlank
 	@Size(min = 4, max = 15)
