+++ src/main/java/com/example/forge/models/entities/Role.java
@@ -5,17 +5,15 @@ import com.example.forge.models.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
+@Table(name = "role")
 @Getter
 @Setter
 @NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "role")
 public class Role extends BaseEntity {
   @Column(unique = true)
   private String name;
