+++ src/main/java/com/example/forge/models/entities/User.java
@@ -15,17 +15,15 @@ import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@NoArgsConstructor
-@AllArgsConstructor
 @Table(name="user")
 @Getter
 @Setter
+@NoArgsConstructor
 public class User extends BaseEntity {
   @NotNull
   @NotBlank
