+++ src/main/java/com/example/forge/repositories/StudentRepository.java
@@ -2,7 +2,6 @@ package com.example.forge.repositories;
 
 import org.springframework.stereotype.Component;
 
-import com.example.forge.baserepository.BaseRepository;
 import com.example.forge.models.entities.Student;
 
 @Component
