+++ src/main/java/com/example/forge/repositories/TeacherRepository.java
@@ -2,9 +2,7 @@ package com.example.forge.repositories;
 
 import org.springframework.stereotype.Component;
 
-import com.example.forge.baserepository.BaseRepository;
 import com.example.forge.models.entities.Teacher;
 
 @Component
-public interface TeacherRepository extends BaseRepository<Teacher>{
-}
+public interface TeacherRepository extends BaseRepository<Teacher>{}
